moved pathway screens to new subfolder
keeping things clean
