+++ App.js
@@ -15,8 +15,8 @@ import ProfileScreen from './screens/ProfileScreen';
 import PathwayScreen from './screens/PathwayScreen';
 import RankingScreen from './screens/RankingScreen';
 import SignOutScreen from './screens/SignOutScreen';
-import PathwayDetailsOneScreen from './screens/PathwayDetailsOneScreen';
-import PathwayDetailsTwoScreen from './screens/PathwayDetailsTwoScreen';
+import PathwayDetailsOneScreen from './screens/Pathways/PathwayDetailsOneScreen';
+import PathwayDetailsTwoScreen from './screens/Pathways/PathwayDetailsTwoScreen';
 
 const Stack = createNativeStackNavigator();
